Replace "float128"
"float128" doesn't work - when I run this code, there is the followring TypeError: data type 'float128' not understood. Therefore, it is necessary to replace it either with float or with longdouble.

+++ estimation/unidimensional/polytomous/grm_mml_eap.py
@@ -97,7 +97,7 @@ def grm_mml_eap(dataset, options=None):
                      'quadrature_bounds': (0.1, 5)}
     hyper_pdf = LatentPDF(hyper_options) 
     hyper_evaluation = np.zeros((hyper_options['quadrature_n'],))
-    base_hyper = (hyper_pdf.weights * hyper_pdf.null_distribution).astype('float128')
+    base_hyper = (hyper_pdf.weights * hyper_pdf.null_distribution).astype(float)
     linear_hyper = base_hyper * hyper_pdf.quadrature_locations
 
     for iteration in range(options['max_iteration']):
